+++ src/main/java/edu/eci/cvds/samples/services/ServiciosAlquiler.java
@@ -19,7 +19,7 @@ public interface ServiciosAlquiler {
      * identificado con 'idcliente'
      * @throws ExcepcionServiciosAlquiler si el cliente no esta registrado
      */
-    public abstract List<ItemRentado> consultarItemsCliente(long idcliente) throws ExcepcionServiciosAlquiler;
+    public abstract List<ItemRentado> consultarItemsCliente(int idcliente) throws ExcepcionServiciosAlquiler;
 
     public abstract List<Cliente> consultarClientes() throws ExcepcionServiciosAlquiler;
